Split `Summary` into precomputed and `Summarizer` structs
Most of the time, we don't need to compute arbitrary percentiles. We
only retained a sorted copy of the summarized sample data support this.
Now we have a `Summarizer`, which retains the sorted data (and can
compute arbitrary percentiles), and change the `Summary` to be a fixed
struct of precomputed summary values.