Use consistent interface for sample data in `LinearRegression`
We move towards a common interface where all sample data is represented
as a slice of sampled _items_. Here, the items are pairs of values taken
from predictor and responder random variables.

Besides the arguable conceptual improvement, this has practical benefits.
The parameter type itself prevents the mismatched length error that we
previously had to check for. We also expect that our typical data sources
will be either loops that range over the predictor variable (so it is
easy to collect pairs), or text formats like TSV, where each line encodes
a single point in the sample, and so the new type is easier to use.