Avoid potential loss of blank space
When rendering with outliers, the columns at which plot chars appeared
were bound by the outer fences, whose columns would obtain the min and
max column values available to the figure. We would never expect to have
extra whitespace outside of the boxplot itself.

If the outer fences instead indicate extreme non-outlier values, the
rows might have blank space that we want to preserve when joining. An
example is if the mean marker is skewed to the left of the left outer
fence (whisker): then there should be blank spaces in the 1st and 3rd
rows, to the left of the top and bottom of the whisker.

By setting the default row cell value to be a single space, joining each
cell on the empty string accounts for such cases when rendering the rows.