Fix deprecation warnings (rust 1.51)
For Error, drop the deprecated description() method and move the logic
into the Display impl.

In bin.rs, use explicit 'dyn' annotation for Boxes of traits.  This is
functionally identical, but the older syntax is deprecated.

In the KAT support logic, the format string of an assert!() must be a
string literal in Rust 2021.  As the variable we were using for the
format string was itself a formatted string, we can just simplify by
doing the formatting in assert!() itself.